% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dados_brutos_telemetria_manaus.R,
%   R/dados_telemetria_manaus.R
\docType{data}
\name{dados_brutos_telemetria_manaus}
\alias{dados_brutos_telemetria_manaus}
\alias{dados_telemetria_manaus}
\title{Dados Brutos da Telemetria 14990000 de Manaus}
\format{
A data frame with 270677 rows and 5 variables

A data frame with 270677 rows and 5 variables
}
\usage{
data(dados_brutos_telemetria_manaus)

data(dados_brutos_telemetria_manaus)
}
\description{
Um banco de dados contendo mais de 270.000 informações.
As variáveis são:

Um banco de dados contendo mais de 270.000 informações. Pegando os dados brutos e
trabalhado da seguinte maneira.
}
\details{
\itemize{
\item CodEstacao. Código da Estação
\item DataHora. Data e Hora do Dado
\item Vazao. Quantidade de Vazão no período
\item Nivel. Nível da Água no momento da medição
\item Chuva. Nível de água captada pela chuva

}

dados_brutos_telemetria_manaus \%>\%
janitor::clean_names() \%>\%
dplyr::na_if('')\%>\%
dplyr::mutate(dplyr::across(c(nivel,chuva,vazao),as.numeric),
data_hora = stringr::str_squish(data_hora) \%>\%
lubridate::as_datetime())

As variáveis são:

\itemize{
\item cod_estacao. Código da Estação
\item data_hora. Data e Hora do Dado
\item vazao. Quantidade de Vazão no período
\item nivel. Nível da Água no momento da medição
\item chuva. Nível de água captada pela chuva

}
}
\keyword{datasets}
