% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historico_bruto_total_manaus.R
\docType{data}
\name{historico_bruto_total_manaus}
\alias{historico_bruto_total_manaus}
\title{Série Histórica Bruta da Telemetria 14990000 de Manaus}
\format{
A data frame with 2854 rows and 78 variables
}
\usage{
data(historico_bruto_total_manaus)
}
\description{
Um banco de dados contendo mais de 270.000 informações. Pegando os dados brutos e
trabalhado da seguinte maneira.
}
\keyword{datasets}
