% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pegar_dados.R
\name{pegar_dados}
\alias{pegar_dados}
\title{Pegar os dados da estação telemétrica}
\usage{
pegar_dados(
  nome_estacao,
  data_inicio = format(floor_date(Sys.Date(), unit = "years"), "\%d/\%m/\%Y"),
  data_fim = format(Sys.Date(), "\%d/\%m/\%Y")
)
}
\arguments{
\item{nome_estacao}{Código da Estação Telemétrica}

\item{data_inicio}{Data com o formato DD/MM/AAAA}

\item{data_fim}{Data com o formato DD/MM/AAAA}
}
\value{
Um tibble com dados telemétricos, puxando os dados do começo do ano vigente até o dia de hoje.
}
\description{
Pegar os dados da estação telemétrica
}
\examples{
pegar_dados(14990000,'01/01/2021','10/01/2021')
}
