% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pegar_historico.R
\name{pegar_historico}
\alias{pegar_historico}
\title{Pegar histórico da série telemétrica da estação}
\usage{
pegar_historico(
  cod_estacao,
  data_inicio = "",
  data_fim = "",
  tipo_dados,
  nivel_consistencia
)
}
\arguments{
\item{cod_estacao}{Código da Estação Telemétrica}

\item{data_inicio}{Data com o formato DD/MM/AAAA}

\item{data_fim}{Data com o formato DD/MM/AAAA}

\item{nivel_consistencia}{1-Bruto ou 2-Consistido}

\item{tipo_dados1-Cotas}{Digitar 1-Cotas, 2-Chuvas ou 3-Vazões}
}
\value{
Um tibble com a série histórica.
}
\description{
Pegar histórico da série telemétrica da estação
}
\examples{
pegar_historico('14990000','01/01/2000','31/01/2000',1,1)

}
